fix: Resolve BCS serialization error for transaction submission
- Wrapped functionality the Aptos object for a uniform API
- Added sign and submit functionality into a single transaction method.
- Provided a Kotlin idiomatic API for carrying out these transactions.

This commit addresses the issue where BCS serialisation outputs where false leading to failed transactions. By encapsulating the transaction logic within the Aptos object, we now have a cleaner, more modular codebase. The new signAndSubmitTransaction method simplifies the transaction process by combining signing and submitting into a single step. Additionally, the API has been refined to be more idiomatic to Kotlin, enhancing code readability and usability.

+++ lib/src/commonMain/kotlin/xyz/mcxross/kaptos/api/Account.kt
@@ -150,7 +150,7 @@ class Account(override val config: AptosConfig) : Account {
   override suspend fun getAccountAPTAmount(
     accountAddress: AccountAddressInput,
     minimumLedgerVersion: Long?,
-  ): Option<Int> =
+  ): Option<Long> =
     getAccountCoinAmount(accountAddress, MoveValue.MoveStructId(APTOS_COIN), minimumLedgerVersion)
 
   /**
@@ -165,5 +165,5 @@ class Account(override val config: AptosConfig) : Account {
     accountAddress: AccountAddressInput,
     coinType: MoveValue.MoveStructId,
     minimumLedgerVersion: Long?,
-  ): Option<Int> = getAccountCoinAmount(config, accountAddress, coinType, minimumLedgerVersion)
+  ): Option<Long> = getAccountCoinAmount(config, accountAddress, coinType, minimumLedgerVersion)
 }
